move the settings button out into the RightHomeContextMenu component
